🔨 Refactor(#337): 승인, 거절 신청자 관리 훅으로 분리
